# TRAAN_Integration_Phase5/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH Disaster Response API",
    "description": "Disaster Response Coordination Platform \u2014 Backend API\n\n**SIH 2026** \u00b7 Component D \u2014 Backend Core\n\nThis Swagger UI is the live contract that Android, Dashboard, and AI/Optimizer teams build against. Do not change endpoint paths, request fields, or response fields without coordinating with all affected teams.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/device/register": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Register Device",
        "operationId": "register_device_api_v1_auth_device_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/authority/login": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Authority Login",
        "operationId": "authority_login_api_v1_auth_authority_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorityLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorityLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/authority/refresh": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Authority Refresh",
        "operationId": "authority_refresh_api_v1_auth_authority_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorityRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorityRefreshResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorityLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "AuthorityLoginRequest"
      },
      "AuthorityLoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "user": {
            "$ref": "#/components/schemas/AuthorityResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "AuthorityLoginResponse"
      },
      "AuthorityRefreshRequest": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "AuthorityRefreshRequest"
      },
      "AuthorityRefreshResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "AuthorityRefreshResponse"
      },
      "AuthorityResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/AuthorityRole"
          },
          "agency": {
            "type": "string",
            "title": "Agency"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "role",
          "agency",
          "email"
        ],
        "title": "AuthorityResponse"
      },
      "AuthorityRole": {
        "type": "string",
        "enum": [
          "DDMA",
          "police",
          "fire",
          "ambulance",
          "NDRF",
          "SDRF",
          "admin"
        ],
        "title": "AuthorityRole"
      },
...
      "DeviceRegisterRequest": {
        "properties": {
          "device_model": {
            "type": "string",
            "title": "Device Model"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          }
        },
        "type": "object",
        "required": [
          "device_model",
          "app_version"
        ],
        "title": "DeviceRegisterRequest"
      },
      "DeviceRegisterResponse": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "device_jwt": {
            "type": "string",
            "title": "Device Jwt"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "device_jwt"
        ],
        "title": "DeviceRegisterResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}